move ut for waitForSnapshotCopy to azuredisk_test.go
